refactor: move components to shared folder
