--- src/composable/composableItem.ts
@@ -1,1 +1,0 @@
-export {};
