+++ src/pages/Profile/MyProfile/components/EditName.vue
@@ -26,7 +26,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 import EditField from '../../components/EditField.vue';
-import HeaderBar from '@/components/HeaderBar.vue';
+import HeaderBar from '@/shared/components/HeaderBar.vue';
 
 const name = ref('');
