+++ src/pages/Profile/MyVehicle/components/VehicleEdit/EditModel.vue
@@ -26,7 +26,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 import EditField from '@/pages/Profile/components/EditField.vue';
-import HeaderBar from '@/components/HeaderBar.vue';
+import HeaderBar from '@/shared/components/HeaderBar.vue';
 
 const model = ref('');
