+++ src/pages/Profile/MyVehicle/components/VehicleEdit/EditYear.vue
@@ -21,7 +21,7 @@
 <script setup lang="ts">
 import { ref } from 'vue';
 import EditField from '@/pages/Profile/components/EditField.vue';
-import HeaderBar from '@/components/HeaderBar.vue';
+import HeaderBar from '@/shared/components/HeaderBar.vue';
 
 const year = ref('');
