+++ src/pages/Profile/MyVehicle/components/VehicleEdit/VehicleEdit.vue
@@ -46,10 +46,10 @@
 import { computed, ref } from 'vue';
 import { RouterLink } from 'vue-router';
 import { type vehicleEditRoutes } from '../../types';
-import ToggleButton from '@/components/ToggleButton.vue';
+import ToggleButton from '@/shared/components/ToggleButton.vue';
 import MaintenanceItem, {
   type MaintenanceItemProps,
-} from '@/components/MaintenanceItem.vue';
+} from '@/shared/components/MaintenanceItem.vue';
 
 const maintenanceNotificationsEnabled = ref(true);
 const maintenances = ref<MaintenanceItemProps[]>([
