+++ src/pages/Profile/ProfilePage.vue
@@ -36,7 +36,7 @@ import { ref } from 'vue';
 import MyProfile from '@/pages/Profile/MyProfile/MyProfile.vue';
 import MyVehicle from '@/pages/Profile/MyVehicle/MyVehicle.vue';
 import picProfile from '@/assets/picProfile.svg';
-import AppHeader from '@/components/AppHeader.vue';
+import AppHeader from '@/shared/components/AppHeader.vue';
 
 const activeTab = ref<'perfil' | 'veiculo'>('perfil');
 const user = ref({
