+++ src/pages/RegistrationMaintenance/AlignmentBalancing.vue
@@ -152,11 +152,11 @@
 
 <script setup lang="ts">
 import { ref, computed } from 'vue';
-import ButtonComponent from '@/components/ButtonComponent.vue';
-import HeaderBar from '@/components/HeaderBar.vue';
+import ButtonComponent from '@/shared/components/ButtonComponent.vue';
+import HeaderBar from '@/shared/components/HeaderBar.vue';
 // import SelectVehicle from '@/components/SelectVehiclePlate.vue';
-import ModalGenerico from '@/components/ModalGenerico.vue';
-import ModalPositive from '@/components/ModalSucess.vue';
+import ModalGenerico from '@/shared/components/ModalGenerico.vue';
+import ModalPositive from '@/shared/components/ModalSucess.vue';
 import helpIcon from '@/assets/helpIcon.svg';
 import alignmentIcon from '@/assets/Alinham.svg';
 import { date as quasarDate } from 'quasar';
