+++ src/pages/SignUp/SignUpStep2.vue
@@ -36,12 +36,12 @@
 import { ref } from 'vue';
 import { useRouter } from 'vue-router';
 import { useRegisterStore } from '@/store/registerStore';
-import SignInUpFooter from '@/components/SignInUpFooter.vue';
-import SignUpTitleStepper from '@/components/SignUpTitleStepper.vue';
-import ButtonComponent from '@/components/ButtonComponent.vue';
-import SignUpPasswordInput from '@/components/SignUpPasswordInput.vue';
-import PasswordChecker from '@/components/PasswordChecker.vue';
-import HeaderBarWithInfo from '@/components/HeaderBarWithInfo.vue';
+import SignInUpFooter from '@/shared/components/SignInUpFooter.vue';
+import SignUpTitleStepper from '@/shared/components/SignUpTitleStepper.vue';
+import ButtonComponent from '@/shared/components/ButtonComponent.vue';
+import SignUpPasswordInput from '@/shared/components/SignUpPasswordInput.vue';
+import PasswordChecker from '@/shared/components/PasswordChecker.vue';
+import HeaderBarWithInfo from '@/shared/components/HeaderBarWithInfo.vue';
 
 const store = useRegisterStore();
 const router = useRouter();
