+++ src/shared/components/AppFooter.vue
