+++ src/shared/components/CustomToggle.vue
