+++ src/shared/components/InputPassword.vue
