+++ src/shared/components/PasswordChecker.vue
