+++ src/shared/components/SelectVehiclePlate.vue
