+++ src/shared/components/SignInUpFooter.vue
