+++ src/shared/components/SignUpPasswordInput.vue
