+++ src/shared/components/SignUpTitleStepper.vue
